move s2 tests to separate folder
