+++ orthopy/s2/__init__.py
@@ -1,5 +1,6 @@
 from . import yu
 from . import zernike
+from . import zernike2
 from .tools import plot, show
 
-__all__ = ["yu", "zernike", "show", "plot"]
+__all__ = ["yu", "zernike", "zernike2", "show", "plot"]
